requirements.txt
langchain_core==1.2.14
langchain_google_genai==4.2.1
langchain_mcp_adapters==0.2.1
langgraph==1.0.9
mcp==1.26.0
python-dotenv==1.2.1
